Add multiple rollback tests
These tests perform multiple rollbacks on the host to verify that the
mechanism works correctly.  After each rollback, the results of
'rpm-ostree status' are inspected to make sure the expected deployment
is in the right place.

There are two versions of the test: one just does a number of rollbacks
and the other does a rollback/reboot combination.