Add support for testing new tree via rebase
This changes introduces the ability to perform the smoketest on a tree
that was delivered via `rpm-ostree rebase`.  In doing so, a number of
tags were introduced for some tasks in order to limit which deliver
mechanism can be used.

For example, when testing a new tree delivered via `rpm-ostree upgrade`,
the test should not run the `rpm-ostree rebase` command (and associated
tasks).  This is accomplished through the use of `--skip-tags
rpm_ostree_rebase` when running the playbook.  A similar exclusion
strategy is used when testing a new tree delivered via `rpm-ostree
rebase`.

This does increase the complexity when running the playbook, but also
increases flexibility.

It would probably be worthwhile to investigate how to simplify this test
by divorcing the choice of delivery (`upgrade` vs. `rebase`) from the
test itself.