improved-sanity-test: Add tests for docker operations (#42)
This brings in some simple docker tests to the sanity suite.  There
are tests performed pre and post upgrade of the system.

Before the upgrade happens, a base image is pulled to be used as part
of the build process of a layered image.  The layered image is built,
run, and then the resulting container is removed.

After the upgrade, the layered image that was built is ran again, the
resulting container removed, and then the layered image is removed.

As part of this change, I also pinned the playbook to Ansible 2.1 as
it seems pretty stable and has been working well thusfar.