improved-sanity-test: Add tests for atomic cli (#45)
* improved-sanity-test: Add tests for atomic cli

This brings in some simple atomic cli tests to the sanity suite. The
subcommands include pull, run, stop, install, uninstall, host status,
and scan.  The same set of tests are run pre and post upgrade.

The improved-sanity-tests playbook is taking longer to run due to
pull large container pulls and pulls from docker.io.  To reduce
overall runtime the atomic tests, the following choices were made:

- Use small rsyslog and busybox container for pull, install, uninstall
- Make atomic scan scan its own openscap image
- Leverage the docker container built in #42 by running atomic run
   and stop between the docker build and docker run commands
- Place cockpit_verify role earlier after the reboot to avoid pulling
   cockpit image twice

* Remove cockpit_run_verify test and role

The cockpit container is now being tested during the testing of the
atomic install, run, and uninstall in atomic_installation_verify
so it was removed.