Use static hostname for 3-way merge check (#55)
* Use static hostname for 3-way merge check

The ansible_hostname fact was changing after reboots in the test
environment causing the improved sanity tests to fail randomly.  A
similar issue occurred in issue #38 with the ansible_fqdn fact.
Using a static hostname still verifies the 3-way merge  after an
upgrade without dealing with issues related to the test environment.

* Use variable for hostname

Use variable for the hostname so we can easily change it.