Add package layering test (#63)
* Add package layering test

Because roles and tasks cannot be be combined in Ansible, roles were created
for positive scenarios in package install and uninstall and its verifications
only.

Negative tests and tests with unexpected output are done afterwards in
the tasks portion of the playbook. Unfortunately there is a lot of repeated
code to validate output in the task section because we cannot combine roles.

* Use include for non positive testing

Since roles could not be combined with regular tasks for the non positive
testing, include was used instead.

Also added a rpmdb check for install and uninstall.

* Make each test a separate play

Reorganizing the playbook into separate plays to facilitate viewing
and triaging of tests.  Each playbook was given a tag so a user can
choose to run specific test(s).

There is still an issue with Ansible ignoring roles that are called
multiple times so some reboots were tagged with different names.