Add admin unlock tests (#69)
* Add admin unlock tests

Adding ostree admin unlock tests.  The strategy to testing admin
unlock is to install packages from the EPEL repo.  I scraped the
HTML for the package named under the assumption that the packages
could and get a new name with the version.

Roles were used for most of the positive testing but negative testing
was done as tasks.  Since roles and tasks cannot be combined, include
was used to re-use roles or to uses roles within roles.

* fixup! Add admin unlock tests

* fixup! Add admin unlock tests

* fixup! Add admin unlock tests