reboot: provide ability to check for started services after reboot (#88)
* reboot: provide ability to check for started services after reboot

Occasionally, we have seen instances where the `docker` daemon was not
starting quickly enough after a reboot, causing subsequent tasks to
fail.  This change introduces the ability to specify a list of
services to check when rebooting the host.

It's slightly messy because we have a `reboot` role and the
`ans_reboot.yml` file that can be included.  So I went the
conservative route and made sure there were defaults specified for
both cases.

I've only changed the use of the `reboot` role in the
`improved-sanity-tests`.  Other places in that suite or in other tests
that reboot the host have not been changed.

* fixup! reboot: provide ability to check for started services after reboot