add a test for rhbz 1421867 (#94)
* add a test for rhbz 1421867

Our internal testing found that trying to use `rpm-ostree status` on a
RHELAH host that was not registered via `subscription-manager` would
fail. See [RHBZ
1421867](https://bugzilla.redhat.com/show_bug.cgi?id=1421867)

To test for this, I've re-arranged when we use `rpm-ostree status` and
added a role that removes any subscriptions from a host and drops an
`unconfigured-state` field into the origin file (optionally).

* roles: add support for stopping the rpm-ostreed dameon

Sometimes we'd like to make sure the `rpm-ostreed` daemon is stopped
before checking the status of the system.  So I've added an optional
variable to the `atomic_host_status_verify` role that will enable the
daemon to be stopped before the status is checked.

* fixup! add a test for rhbz 1421867