Save failure details to result file (#92)
* Save failure details to result file

Added a callback_plugin that makes the logs more human readable and
also saves down failure details to a file.

A majority of this code is actually from other projects.  The code
that alters the Ansible output is from openshift-ansible which is
derived from an upstream project (in the comments).  I had to override
the baseclass methods to track the failure details.  In order to
preserve the default Ansible logging, I copied the original method
code back into the override method.

All test that want to leverage the callback_plugin need a symlink to
the top callback_plugins directory otherwise the default Ansible logging
will be used.

* fixup! Save failure details to result file