i-s-t: add tag for testing cloud image (#106)
The main outcome of this change is to introduce a tag that allows
users to run the `improved-sanity-test` playbook against a freshly
minted cloud image.  This causes portions of the playbook to be
skipped, most notably the upgraded and rollback tests.  The idea for
this addition came from the observation that Fedora runs tests via
`autocloud` against each new Atomic Host cloud image that produced
every night.

To use this new tag, a user would invoke the playbook with the `--tag
cloud_image` value:

`$ ansible-playbook -i inventory tests/improved-sanity-test/main.yml --tag cloud_image`

Currently, the order of tests/operations for this new tag should look
like this:

- compare `atomic host status` vs `rpm-ostree status`
- verify SELinux labels
- modify SELinux file contexts
- check `docker pull`, `run`, `remove`
- use pkg-layering to install a package
- use `ostree admin unlock` to install an RPM
- verify `/tmp` permissions
- verify the rpmdb is usable
- add users, modify files in `/etc/`, drop files in `/var`
- test `atomic pull`, `atomic scan`
- `docker build` and `docker run` image
- `atomic install` cockpit container
- cleanup any containers + images

As part of this effort, I re-ordered when the `osname_set_fact` role
was run and used `always_run: yes` with that role.