Update tests to run on Ansible 2.2 (#109)
* Update tests to run on Ansible 2.2

This check-in makes the necessary changes in the current supported
tests to run on Ansible 2.2.  The most notable change was in the
callback_plugin which is not backwards compatible with older versions
of Ansible.  Because of this, I made sure there was a check for
Ansible version 2.2 in every supported test.

There were a few roles that used the docker module that needed to be
modified to work with Ansible 2.2.  Other than that it was mostly
documentation changes.

* fixup! Update tests to run on Ansible 2.2

* fixup! Update tests to run on Ansible 2.2