semanage_add_verify: fix to run on all platforms (#110)
* semanage_add_verify: fix to run on all platforms

Previously we were trying to change the `fcontext` object type on a
specific directory based on if there was a `kubelet` package
installed.  This was incorrect, as I don't think any platform/stream
had that package installed.

This re-tools the role to modify an known object type to a new value
and compares the result at the end.  No uncessary checks and should
work on all streams.

Fixes #99

* roles: split semanage actions into separate roles

In order to be more flexible when used in the main workhorse that is
the 'improved-sanity-tests', I've split the actions around the
'semanage' functionality into two roles.  Now the modification of the
SELinux file context happens separately from the verification.