Cleanup unused files and directories (#117)
* Cleanup unused files and directories

The original creation of this repo did not have a clear direction or
set of guidelines on how we wanted to structure the playbooks.  As
such, there were a number of extra directories and files that were
created as we found our way.

We are in a better place now and have a better sense of how we want
the repo to look.  This is the first of a few commits that will
cleanup the repo and hopefully make it less of a mess.

* fixup! Cleanup unused files and directories