Correct some conditional statements (#127)
While re-testing the various playbooks after I had cleaned up the
repo, I found a few places where our conditionals were incorrect.

This fixes these errors and sneaks in a typo fix as well.

NB: this is the last of three commits that cleanup the repo.  The
other two commits should be merged first before we get to this one.