Assume user could be root when checking privileges (#137)
* Assume user could be root when checking privileges

In #136 a test was added to check that a non-root user could run rpm-ostree
status.  There was an assumption made that the user running the test was
not a root user.  It is bad to assume.  This check-in will now run rpm-ostree
status as bin using runuser.

To facilitate additional privilege checks, two new roles were added that will
verify that commands can/cannot be run by a specific user.

* fixup! Assume user could be root when checking privileges

* fixup! Assume user could be root when checking privileges