Add additional tests for package layering (#135)
* Add additional tests for package layering

The functionality for package layering has changed slightly so the
test need to be updated.  Two tests were added to verify the new
fucntionality.  The first test verifies that packages that are
being layered is added to a list of requested packages even if it is
already in the deployment.  The second test is that rpm-ostree should
error out when attempting to layer a package that already exists.

There was also a change in the rpm-ostree status output.  When there
are no packages are layered, there will be no "packages" listed in
the rpm-ostree status output. This caused an error when setting facts
in  rpm_ostree_uninstall_verify role.  To get around this, I only
set the fact to {} if it is undefined or to the list of packages if
it is defined.

* fixup! Add additional tests for package layering