Collect journal on failure (#144)
* Collect journal on failure

This commit will collect the journal for the improved sanity test and
save it locally.  It also enables running other tasks on failure.

This is a giant workaround since Ansible does not have direct support to
run post execution tasks on failure..  Ansible's only mechanism for running
tasks after a play completes is a handler.  The problem with a handler
is that it will not run on failure without the force_handlers set to true.
If force_handlers is set to true it will run after every block of
pre_tasks, post_task, roles, and plays (not desirable).  The way to
workaround that is to conditionalize the tasks to only run when there
is a failure; however, Ansible does not have a direct variable that holds
the pass/fail status of a playbook.

In order to get the fail status, I leveraged the v2_runner_on_failed
method in the callback plugin that is called when Ansible fails.
An environment variable with the fail status of the result. Now the
handler tasks can be conditionalized to run on failure based on the
environment variable.

The handler can notify tasks individually by name or tasks can listen for
a notification.  Since multiple tasks need to be run, I opted to have
all tasks that need to be run together to listen for the same notification.
Notified tasks are run in the order they coded and not by the time they
are notified.

* fixup! Collect journal on failure