Support passing extra options to sub. man. (#153)
* Support passing extra options to sub. man.

Sometimes additonal options are needed during registration, such as
``--force`` or ``--org``.  Support this with an optional 'extra' key in
the CSV file.  When this key is not present, treat it as an empty
string to the sub. man. command-line.

Signed-off-by: Chris Evich <cevich@redhat.com>

* fixup! Support passing extra options to sub. man.

Add back in use of --force (accidentally dropped by me)