tests: OpenShift Ansible Installer sanity test (#162)
* tests: OpenShift Ansible Installer sanity test

This is a 'meta' playbook that borrowed heavily from the work done in
'ansible-ansible-openshift-ansible' [0].  It will install an OpenShift
Origin cluster on a single host via the openshift/openshfit-ansible
installer.

This playbook can be invoked as you would for any other tests in this
repo, by supplying an inventory file or hostname/IP address to
`ansible-playbook`.  The host from the inventory is used to template
out a second inventory file which is used when calling
`ansible-playbook` on the openshift/openshift-ansible playbook. This,
in theory, allows users to maintain the same workflow that is familiar
to running other tests in this repo.

There are variables defined in the playbook that control the version
of OpenShift Origin to be installed, the `git` tag of the
openshift/openshift-ansible repo to use, the host where the cluster
should be installed, and the Ansible user to use when installing the
cluster.  These can all be overridden via CLI parameters.

This lacks a couple of things that could be added later:
  - additional checks to determine health of cluster
  - cleanup/uninstall of the cluster
  - maybe deploying a project?

* openshift-ansible-testing: make it work with CentOS + RHEL AH

- added symlinks to callback_plugins + roles
- introduced new vars to handle public vs private ip addresses
- added conditional to skip memory check
- added conditional to use Python3 for Fedora only
- added README.md

* fixup! openshift-ansible-testing: make it work with CentOS + RHEL AH

* fixup! openshift-ansible-testing: make it work with CentOS + RHEL AH