Simplify rpm install verification in improved sanity test (#177)
* Simplify rpm install verification in improved sanity test

Addresses issue #175.  The previous method of verifying rpm installs
relied scraping HTML to determine the filename for a RPM before it
could install it.  There was also a single point of failure pointing
to dl.fedoraproject.org instead of the mirror site.

This PR simplifies the verification in the improved sanity test by
using the epel-release package in the mirror site which always has
the same URL.

* fixup! Simplify rpm install verification in improved sanity test

* Generalize rpm_install role

Instead of having a role specifically to install packages from EPEL,
it was suggested to have a generic role that installs any package if
given a url.  In order to preserve the functionality of the origin role,
a separate role (epel_rpm_url) was created that returns the full url of
a package after scraping HTML.  These two roles replace the current
unctionality of rpm_install.

After the role changes, the improved-sanity-test and admin-unlock tests
needed to be modified to work with the new roles.