Update handling of /etc/sysconfig/docker for docker-latest (#197)
* Update handling of /etc/sysconfig/docker for docker-latest

The more recent versions (1.13, I think) of `docker-latest`
require additional lines to be modified in `/etc/sysconfig/docker` for
everything to play happy together.

Closes #192

* docker_build_httpd: add retry logic

As part of the effort to enable retries on network operations, this
turns on retries when doing a `docker build`.  This is because the
operation has to fetch metatdata and RPM data from the network as part
of the build, so might as well try a few times in case the network
sucks.