add support for /etc/containers/registries.conf (#206)
Supported is needed for an additional  scenario now for adding
registries and insecure registries.  The current way to add
registries is in /etc/sysconfig/docker.  The new way is to add
registries to /etc/containers/registries.conf which uses yaml
format.  If /etc/containers/registries.conf exists, it the
registry will be added in the file, otherwise it will be added
in /etc/sysconfig/docker.